+++ src/main/java/gr/knowledge/internship/banksystem/dto/CodedTypeDTO.java
@@ -1,6 +1,5 @@
 package gr.knowledge.internship.banksystem.dto;
 
-
 import gr.knowledge.internship.banksystem.entity.CodedType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,16 +8,14 @@ import lombok.Setter;
 
 import java.io.Serializable;
 
+@Setter
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Getter
-@Setter
-public class PhoneDTO implements Serializable {
+public class CodedTypeDTO implements Serializable {
+
     private Long id;
-
-    private CodedTypeDto phoneType;
-
-    private String phoneNumber;
-
-    private String registryAfm;
+    private String code;
+    private String description;
+    private CodedType codedType;
 }
