+++ src/main/java/gr/knowledge/internship/banksystem/dto/PhoneDTO.java
@@ -1,6 +1,6 @@
 package gr.knowledge.internship.banksystem.dto;
 
-import gr.knowledge.internship.banksystem.entity.CodedType;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,14 +8,16 @@ import lombok.Setter;
 
 import java.io.Serializable;
 
-@Setter
-@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CodedTypeDto implements Serializable {
-
+@Getter
+@Setter
+public class PhoneDTO implements Serializable {
     private Long id;
-    private String code;
-    private String description;
-    private CodedType codedType;
+
+    private CodedTypeDTO phoneType;
+
+    private String phoneNumber;
+
+    private String registryAfm;
 }
