+++ src/main/java/gr/knowledge/internship/banksystem/mapper/CodedTypeMapper.java
@@ -1,11 +1,7 @@
 package gr.knowledge.internship.banksystem.mapper;
 
-import gr.knowledge.internship.banksystem.dto.CodedTypeDto;
+import gr.knowledge.internship.banksystem.dto.CodedTypeDTO;
 import gr.knowledge.internship.banksystem.entity.CodedType;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -19,19 +15,19 @@ public class CodedTypeMapper {
     @Autowired
     private ModelMapper modelMapper;
 
-    public CodedTypeDto toDto(CodedType codedType) {
-        return modelMapper.map(codedType, CodedTypeDto.class);
+    public CodedTypeDTO toDto(CodedType codedType) {
+        return modelMapper.map(codedType, CodedTypeDTO.class);
     }
 
-    public CodedType toEntity(CodedTypeDto codedTypeDto) {
+    public CodedType toEntity(CodedTypeDTO codedTypeDto) {
         return modelMapper.map(codedTypeDto, CodedType.class);
     }
 
-    public List<CodedType> toEntityList(List<CodedTypeDto> codedTypeDtoList) {
+    public List<CodedType> toEntityList(List<CodedTypeDTO> codedTypeDtoList) {
         return codedTypeDtoList.stream().map(this::toEntity).collect(Collectors.toList());
     }
 
-    public List<CodedTypeDto> toDtoList(List<CodedType> codedTypeList) {
+    public List<CodedTypeDTO> toDtoList(List<CodedType> codedTypeList) {
         return codedTypeList.stream().map(this::toDto).collect(Collectors.toList());
     }
 }
