+++ src/main/java/gr/knowledge/internship/banksystem/repository/CodedTypeRepository.java
